# Allows email address domains to be IP addresses