Sorry I didn't make an issue first. Figured it's a small change, so I'd just offer a solution whether it's acceptable or not.

References:
- https://learn.microsoft.com/en-us/archive/blogs/testing123/email-address-test-cases
- https://github.com/validatorjs/validator.js/issues/800

This doesn't make sure the IP address is valid. It only makes sure that it's not obviously invalid.