@kibertoad 2022-09-29
can you benchmark perf impact of this change on plain emails?