@kibertoad 2022-09-29
I think validator.js approach of making this optional generally makes more sense, especially if this makes the hot path regex slower