@Tyler-Murphy 2022-09-30
Sure, added to `benchmarks/string.ts` like this:

```ts
const emailSchema = z.string().email()

...

  .add("email", () => {
    emailSchema.parse(`someone@example.com`)
  })
```

Without IP support: 7,224,694 ops/sec ±0.59% (96 runs sampled)
With IP support: 7,175,461 ops/sec ±0.62% (92 runs sampled)

I also tried with 1000 random email addresses from https://www.randomlists.com/email-addresses?qty=1000 where each one was checked in a `for .. of` loop.

Without IP support: 7,459 ops/sec ±0.39% (97 runs sampled)
With IP support: 7,584 ops/sec ±0.56% (97 runs sampled)

And finally, I replaced two of the 1000 random email addresses with email addresses that use IP addresses, still using the regex with IP support: 7,663 ops/sec ±0.41% (97 runs sampled)

The speed is so similar that I don't think it's worth making it optional.