@colinhacks 2024-04-07
Thanks for this (now very old) PR! The final word on email regexes landed here: https://github.com/colinhacks/zod/pull/2157 Further discussion can happen in new issues/PRs. 👍